Robust Point Count (Image Menu)
Allows identification of up to 9 pre-set user phases. All data stored in text-file afterwards. Allows user to save images of frame at the end of a specific frame analysis (with naming of frame corresponding to its location). Incorporates a pause prior to moving to next frame to allow analysis cross-channels.